Move trainee dashboard tests to a new file
From test_landing_page.py to test_trainee_dashboard.py.
