+++ workshops/test/test_landing_page.py
@@ -48,22 +48,6 @@ class TestAdminDashboard(TestBase):
         self.assertNotIn(completed, response.context['unpublished_events'])
 
 
-class TestTraineeDashboard(TestBase):
-    """ Tests for trainee dashboard. """
-    def setUp(self):
-        self.user = Person.objects.create_user(
-            username='user', personal='', family='',
-            email='user@example.org', password='pass')
-        self.client.login(username='user', password='pass')
-
-    def test_dashboard_loads(self):
-        rv = self.client.get(reverse('trainee-dashboard'))
-        self.assertEqual(rv.status_code, 200)
-        content = rv.content.decode('utf-8')
-        self.assertIn("Log out", content)
-        self.assertIn("Update your profile", content)
-
-
 class TestDispatch(TestBase):
     """ Test that the right dashboard (trainee or admin dashboard) is displayed
     after logging in."""
